moved IPP under satellites
